Conflicts:
	src/Catty/App-Info.plist

+++ src/Catty/Extension&Delegate&Protocol/Extensions/UIColor+Extension/UIColor+CatrobatUIColorExtensions.m
@@ -45,10 +45,10 @@
 
 + (UIColor*)globalTintColor
 {
-    return [UIColor colorWithRed:(CGFloat)(232.0/255.0) green:146.0f/255.0f blue:6.0f/255.0f alpha:1.0f];
+    return [UIColor colorWithRed:232.0f/255.0f green:146.0f/255.0f blue:6.0f/255.0f alpha:1.0f];
 }
 
-+ (UIColor*) utilityTintColor
++ (UIColor*)utilityTintColor
 {
     return [self lightTextTintColor];
 }
@@ -70,14 +70,13 @@
 
 + (UIColor*)lightTextTintColor
 {
-    return [UIColor colorWithRed:168.0f/255.0f green:223.0f/255.0f blue:244/255.0f alpha:1.0f];
+    return [UIColor colorWithRed:168.0f/255.0f green:223.0f/255.0f blue:244.0f/255.0f alpha:1.0f];
 }
 
 + (UIColor*)darkTextTintColor
 {
     return [UIColor colorWithRed:111.0f/255.0f green:142.0f/255.0f blue:155.0f/255.0f alpha:1.0f];
 }
-
 
 + (UIColor*)destructiveTintColor
 {
